requirements.txt
cmake
dlib
face_recognition
pillow
flask
flask_script
opencv-python